+++ Development-CloudComputing/credential-firebase.txt
@@ -1,0 +1,8 @@
+  apiKey: "AIzaSyDZ2MtR_Iqaewx8iBsFo8GvEME5KwL4jFk",
+  authDomain: "signmate-7778d.firebaseapp.com",
+  databaseURL: "https://signmate-7778d-default-rtdb.asia-southeast1.firebasedatabase.app",
+  projectId: "signmate-7778d",
+  storageBucket: "signmate-7778d.appspot.com",
+  messagingSenderId: "215899096912",
+  appId: "1:215899096912:web:e6c811b66b31d286cf8e55",
+  measurementId: "G-Y9D1F6E34G"
