+++ app/src/main/AndroidManifest.xml
@@ -13,28 +13,28 @@
         android:theme="@style/Theme.SignMateC241PS262"
         tools:targetApi="31">
         <activity
-            android:name=".PlayNowTF"
+            android:name=".ui.play.PlayNowTF"
             android:exported="false" />
         <activity
             android:name=".PlayNow_number"
             android:exported="false" />
         <activity
-            android:name=".PlayNowAlphabet"
+            android:name=".ui.play.PlayNowAlphabet"
             android:exported="false" />
         <activity
-            android:name=".QuizTFActivity"
+            android:name=".ui.quiz.QuizTFActivity"
             android:exported="false" />
         <activity
-            android:name=".QuizNumberActivity"
+            android:name=".ui.quiz.QuizNumberActivity"
             android:exported="false" />
         <activity
-            android:name=".QuizAlphabetActivity"
+            android:name=".ui.quiz.QuizAlphabetActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".login.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.dashboard.DashboardActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
