--- app/src/main/java/com/capstone/signmate_c241_ps262/OpeningActivity.kt
@@ -1,4 +1,0 @@
-package com.capstone.signmate_c241_ps262
-
-class OpeningActivity {
-}
