+++ app/src/main/java/com/capstone/signmate_c241_ps262/ui/dashboard/DashboardActivity.kt
@@ -1,16 +1,17 @@
-package com.capstone.signmate_c241_ps262
+package com.capstone.signmate_c241_ps262.ui.dashboard
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.capstone.signmate_c241_ps262.R
 
-class PlayNowAlphabet : AppCompatActivity() {
+class DashboardActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_play_now_alphabet)
+        setContentView(R.layout.activity_dashboard)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
