+++ app/src/main/java/com/capstone/signmate_c241_ps262/ui/opening/OpeningActivity.kt
@@ -1,0 +1,4 @@
+package com.capstone.signmate_c241_ps262.ui.opening
+
+class OpeningActivity {
+}
