+++ app/src/main/java/com/capstone/signmate_c241_ps262/ui/play/PlayNowTF.kt
@@ -1,10 +1,11 @@
-package com.capstone.signmate_c241_ps262
+package com.capstone.signmate_c241_ps262.ui.play
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.capstone.signmate_c241_ps262.R
 
 class PlayNowTF : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
