+++ app/src/main/res/layout/activity_login.xml
@@ -4,10 +4,10 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    android:background="#F5F7FA"
+    android:background="@color/white_page"
     android:orientation="vertical"
     android:gravity="center"
-    tools:context=".LoginActivity">
+    tools:context=".login.LoginActivity">
 
     <ImageButton
         android:layout_width="66dp"
