+++ app/src/main/res/layout/activity_play_now_alphabet.xml
@@ -3,8 +3,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:backgroundTint="@color/white_tint"
-    tools:context=".PlayNowAlphabet">
+    android:background="@color/white_page"
+    tools:context=".ui.play.PlayNowAlphabet">
 
     <ImageButton
         android:layout_width="66dp"
