+++ app/src/main/res/layout/activity_quiz_alphabet.xml
@@ -3,8 +3,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:backgroundTint="@color/white_tint"
-    tools:context=".QuizAlphabetActivity">
+    android:background="@color/white_page"
+    tools:context=".ui.quiz.QuizAlphabetActivity">
 
     <TextView
         android:id="@+id/tv_quizAndPractice"
