+++ app/src/main/res/layout/activity_quiz_number.xml
@@ -2,7 +2,7 @@
 <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:backgroundTint="@color/white_tint">
+    android:background="@color/white_page">
 
     <TextView
         android:id="@+id/tv_quizAndPractice"
