+++ app/src/main/res/layout/activity_quiz_tfactivity.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:backgroundTint="@color/white_tint"
+    android:background="@color/white_page"
     android:layout_height="match_parent">
 
     <TextView
