+++ app/src/main/res/values/colors.xml
@@ -2,9 +2,13 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="white_tint">#F8FBFF</color>
+    <color name="white_page">#F8FBFF</color>
+    <color name="white_text">#EEF7FF</color>
     <color name="yellow">#FFCC3E</color>
     <color name="red">#E91E63</color>
     <color name="blue">#030047</color>
     <color name="blue_dark">#000427</color>
+    <color name="gray">#D9D9D9</color>
+    <color name="blue_bright">#88D9E1</color>
+    <color name="cream_white">#F3F7EC</color>
 </resources>
