+++ app/src/main/res/values/strings.xml
@@ -12,4 +12,14 @@
     <string name="quiz_number">What number does the sign\nlanguage picture refer\nto?</string>
     <string name="quiz_yes_no">Is this Z?</string>
     <string name="play_now">Play Now</string>
+    <string name="welcome">Hi User | Guest!</string>
+    <string name="manage_profile">Manage Profile</string>
+    <string name="about">About</string>
+    <string name="feedback">Feedback</string>
+    <string name="app_title">SignMate</string>
+    <string name="menu">Menu</string>
+    <string name="menu_1">Sign Language Dictionary</string>
+    <string name="menu_2">Hand Recognition</string>
+    <string name="menu_3">Quiz Practice</string>
+    <string name="our_feature">Our Feature</string>
 </resources>
